fix(gateway): a heartbeat queued behind the ending writes nothing once the turn is sealed
heartbeat() read the seal on the way in and never again, so the gap between
clearing that read and holding progressMutex was enough for the whole ending to
run — CAS to ENDING, closeProgress, message_delta, message_stop — after which the
queued write put a ping past the terminal. Two ways in, and the second needs no
lock at all: the pinger clears the entry read, is descheduled, the terminal
completes, and the pinger then takes an UNCONTENDED mutex and writes. That one
existed in b481124d, where heartbeat was `if (seal != OPEN) return;
pingIfStarted()` — the same check-then-act gap with nothing in it. So the mutex
neither introduced this nor fixed it; it widens the contended path specifically,
because a waiter is resumed exactly as the terminal proceeds to its last frames.

The pinger cannot be assumed gone: TurnOneDrive cancels it in its finally, after
the round returns, so it is live throughout emitTerminal, and Ktor closes the
response later still — a late frame reaches the socket rather than dying on a
closed channel.

Severity was ordering, not corruption: fixed bytes, and the progress port keeps it
out of content_frames_out either way. A ping between message_delta and
message_stop is legal Anthropic grammar; past message_stop it is junk on a
finished stream. What was actually wrong is that heartbeat's own docstring
promises "a turn that has claimed or reached its ending writes nothing more" and
heartbeat was the one verb not honouring it.

Fix mirrors progress(): re-read the seal under the lock. Sufficient because
closeProgress runs after the CAS — acquire after it and you see non-OPEN and write
nothing; acquire before it and your ping lands ahead of message_delta, which is
legal.

Found by splice-astra in review of the pre-existing code, who also chose the
assertion: "no queued heartbeat write after ENDING" pins the invariant, where
"nothing after message_stop" would have pinned a schedule. The barrier is the
test's own write sink — no instrumentation in production code. The regression
holds the lock with a blocked progress line, queues a heartbeat behind it while
the seal is still OPEN, then lets the terminal claim the seal and queue behind
both; an uncontended heartbeat first proves the verb writes at all, so the final
count of one ping cannot pass vacuously.

Deleting the under-lock check makes that two pings — verified red before this
commit and again after the file split, then green restored, both times.

The five pinger tests move to SseEmitterProgressTest: adding this one pushed
SseEmitterTest past detekt's LargeClass ceiling, and the seam is a coherent file
of its own rather than a suppression. Gate green against app-all.jar sha256
9d4a441b; oracle 11/11, walls 0, rules 47/47.
